refactors lines and cmus tags
