+++ rust/lolfm/src/actions/init_db.rs
@@ -1,9 +1,10 @@
 use crate::models::app_config::AppConfig;
 use crate::models::er::Er;
-use crate::repos::cmus_tags_from_shell;
 use crate::repos::db_create_schema;
+use crate::repos::lines_from_cmus_remote;
 use crate::repos::songs_to_db;
 use crate::transformers::cmus_tags_to_songs;
+use crate::transformers::lines_to_cmus_tags;
 
 pub fn run(ω: &AppConfig) -> Result<(), Er> {
   match process(ω) {
@@ -15,7 +16,8 @@ pub fn run(ω: &AppConfig) -> Result<(), Er> {
 fn process(ω: &AppConfig) -> Result<(), Er> {
            ω.db.execute("BEGIN TRANSACTION")?;
            db_create_schema::run(&ω.db)?;
-  let ts = cmus_tags_from_shell::get(&["-C", "save -l -e -"], "file")?;
+  let ls = lines_from_cmus_remote::get(&["-C", "save -l -e -"])?;
+  let ts = lines_to_cmus_tags::run(ls, "file");
   let ss = cmus_tags_to_songs::run(ts);
            songs_to_db::write(&ω.db, ss)
 }
