+++ rust/lolfm/src/actions/process_cmus_event.rs
@@ -2,10 +2,11 @@ use crate::models::app_config::AppConfig;
 use crate::models::er::Er;
 use crate::repos::cmus_events_from_db;
 use crate::repos::cmus_events_to_db;
-use crate::repos::cmus_tags_from_shell;
+use crate::repos::lines_from_cmus_remote;
 use crate::repos::lolfm_events_to_db;
 use crate::transformers::cmus_events_to_lolfm_events;
 use crate::transformers::cmus_tags_to_cmus_events;
+use crate::transformers::lines_to_cmus_tags;
 
 pub fn run(ω: &AppConfig) -> Result<(), Er> {
   match process(ω) {
@@ -16,7 +17,8 @@ pub fn run(ω: &AppConfig) -> Result<(), Er> {
 
 fn process(ω: &AppConfig) -> Result<(), Er> {
            ω.db.execute("BEGIN TRANSACTION")?;
-  let ts = cmus_tags_from_shell::get(&["-Q"], "status")?;
+  let ls = lines_from_cmus_remote::get(&["-Q"])?;
+  let ts = lines_to_cmus_tags::run(ls, "status");
   let cs = cmus_tags_to_cmus_events::run(ts, ω.time);
            cmus_events_to_db::write(&ω.db, cs)?;
   let es = cmus_events_from_db::get(&ω.db)?;
